# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Millitary service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/bullet/add": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add quantity to a Bullet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bullet"
                ],
                "summary": "Add Quantity",
                "parameters": [
                    {
                        "description": "Bullet data",
                        "name": "Bullet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/militaries.BulletAddSub"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Add Successful",
                        "schema": {
                            "$ref": "#/definitions/militaries.Void"
                        }
                    },
                    "500": {
                        "description": "Error while adding quantity",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bullet/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new bullet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bullet"
                ],
                "summary": "Create Bullet",
                "parameters": [
                    {
                        "description": "Bullet Request",
                        "name": "BulletReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/militaries.BulletReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while creating",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bullet/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an existing bullet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bullet"
                ],
                "summary": "Delete Bullet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bullet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while deleting",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bullet/getall": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all bullets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bullet"
                ],
                "summary": "Get All Bullets",
                "parameters": [
                    {
                        "type": "number",
                        "name": "caliber",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "quantity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get All Successful",
                        "schema": {
                            "$ref": "#/definitions/militaries.AllBullets"
                        }
                    },
                    "401": {
                        "description": "Error while getting all",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bullet/getbyid/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get an existing bullet by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bullet"
                ],
                "summary": "Get Bullet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bullet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Successful",
                        "schema": {
                            "$ref": "#/definitions/militaries.Bullet"
                        }
                    },
                    "401": {
                        "description": "Error while getting",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bullet/sub": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Subtract quantity from a Bullet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bullet"
                ],
                "summary": "Subtract Quantity",
                "parameters": [
                    {
                        "description": "Bullet data",
                        "name": "Bullet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/militaries.BulletAddSub"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subtract Successful",
                        "schema": {
                            "$ref": "#/definitions/militaries.Void"
                        }
                    },
                    "500": {
                        "description": "Error while subtracting quantity",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bullet/update/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing bullet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bullet"
                ],
                "summary": "Update Bullet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bullet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Bullet",
                        "name": "Bullet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/militaries.Bullet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while updating",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "militaries.AllBullets": {
            "type": "object",
            "properties": {
                "bullets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/militaries.Bullet"
                    }
                }
            }
        },
...
        "militaries.Bullet": {
            "type": "object",
            "properties": {
                "caliber": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "militaries.BulletAddSub": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "militaries.BulletReq": {
            "type": "object",
            "properties": {
                "caliber": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "militaries.Void": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authourization",
            "in": "header"
        }
    }
}